split copy tests into their own file
